fix(aero-jit-x86): remove redundant casts for clippy
- Drop unnecessary width.bytes() usize cast in Tier-2 interpreter
- Simplify test constants/offset math to avoid unnecessary_cast/identity_op

+++ crates/aero-jit-x86/src/tier2/interp.rs
@@ -151,7 +151,7 @@ fn exec_instr(
             let val = eval_operand(src, values);
             stats.mem_stores += 1;
             bus.write(addr, width, val);
-            env.page_versions.bump_write(addr, width.bytes() as usize);
+            env.page_versions.bump_write(addr, width.bytes());
         }
         Instr::Guard {
             cond,
